+++ .evergreen/run-csfle-tests.sh
@@ -11,6 +11,7 @@ set -o xtrace
 export CSFLE_TLS_CERT_DIR="${DRIVERS_TOOLS}/.evergreen/x509gen"
 
 FEATURE_FLAGS+=("in-use-encryption" "azure-kms")
+CARGO_OPTIONS+=("--ignore-default-filter")
 
 if [[ "$OPENSSL" = true ]]; then
   FEATURE_FLAGS+=("openssl-tls")
@@ -33,11 +34,11 @@ cargo_test test::csfle
 cargo_test test::spec::client_side_encryption
 
 FEATURE_FLAGS+=("aws-auth")
-cargo_test test::csfle::prose::on_demand_aws::success
+cargo_test test::csfle::on_demand_aws::success
 
 # Unset variables for on-demand credential failure tests.
 unset AWS_ACCESS_KEY_ID
 unset AWS_SECRET_ACCESS_KEY
-cargo_test test::csfle::prose::on_demand_aws::failure
+cargo_test test::csfle::on_demand_aws::failure
 
 exit ${CARGO_RESULT}
