+++ src/test/csfle/prose.rs
@@ -9,7 +9,7 @@ use std::{
 };
 
 use futures_util::TryStreamExt;
-use mongocrypt::ctx::{Algorithm, KmsProvider};
+use mongocrypt::ctx::Algorithm;
 use tokio::net::TcpListener;
 
 use crate::{
@@ -65,7 +65,6 @@ use super::{
     validate_roundtrip,
     Result,
     AWS_KMS,
-    AZURE_IMDS_MOCK_PORT,
     DISABLE_CRYPT_SHARED,
     EXTRA_OPTIONS,
     KV_NAMESPACE,
@@ -1449,171 +1448,14 @@ mod decryption_events {
     }
 }
 
-// Prose test 15. On-demand AWS Credentials
-#[cfg(feature = "aws-auth")]
-mod on_demand_aws {
-    use super::*;
-
-    #[tokio::test]
-    async fn failure() -> Result<()> {
-        assert!(std::env::var("AWS_ACCESS_KEY_ID").is_err());
-        assert!(std::env::var("AWS_SECRET_ACCESS_KEY").is_err());
-
-        let ce = ClientEncryption::new(
-            Client::for_test().await.into_client(),
-            KV_NAMESPACE.clone(),
-            [(KmsProvider::aws(), doc! {}, None)],
-        )?;
-        let result = ce
-            .create_data_key(
-                AwsMasterKey::builder()
-                    .region("us-east-1")
-                    .key(
-                        "arn:aws:kms:us-east-1:579766882180:key/\
-                         89fcc2c4-08b0-4bd9-9f25-e30687b580d0",
-                    )
-                    .build(),
-            )
-            .await;
-        assert!(result.is_err(), "Expected error, got {:?}", result);
-
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn success() -> Result<()> {
-        assert!(std::env::var("AWS_ACCESS_KEY_ID").is_ok());
-        assert!(std::env::var("AWS_SECRET_ACCESS_KEY").is_ok());
-
-        let ce = ClientEncryption::new(
-            Client::for_test().await.into_client(),
-            KV_NAMESPACE.clone(),
-            [(KmsProvider::aws(), doc! {}, None)],
-        )?;
-        ce.create_data_key(
-            AwsMasterKey::builder()
-                .region("us-east-1")
-                .key("arn:aws:kms:us-east-1:579766882180:key/89fcc2c4-08b0-4bd9-9f25-e30687b580d0")
-                .build(),
-        )
-        .await?;
-
-        Ok(())
-    }
-}
-
 // TODO RUST-1441: implement prose test 16. Rewrap
-
-// Prose test 17. On-demand GCP Credentials
-#[cfg(feature = "gcp-kms")]
-mod on_demand_gcp {
-    use super::*;
-
-    async fn try_create_data_key() -> crate::error::Result<()> {
-        let util_client = Client::for_test().await.into_client();
-        let client_encryption = ClientEncryption::new(
-            util_client,
-            KV_NAMESPACE.clone(),
-            [(KmsProvider::gcp(), doc! {}, None)],
-        )?;
-
-        client_encryption
-            .create_data_key(
-                GcpMasterKey::builder()
-                    .project_id("devprod-drivers")
-                    .location("global")
-                    .key_ring("key-ring-csfle")
-                    .key_name("key-name-csfle")
-                    .build(),
-            )
-            .await
-            .map(|_| ())
-    }
-
-    #[tokio::test]
-    async fn failure() {
-        let error = try_create_data_key().await.unwrap_err();
-        match *error.kind {
-            ErrorKind::Encryption(e) => {
-                assert!(matches!(e.kind, mongocrypt::error::ErrorKind::Kms));
-                assert!(e.message.unwrap().contains("GCP credentials"));
-            }
-            other => panic!("Expected encryption error, got {:?}", other),
-        }
-    }
-
-    // run in GCP environment
-    #[tokio::test]
-    async fn success_skip_ci() {
-        try_create_data_key().await.unwrap();
-    }
-}
-
-// Prose test 18. Azure IMDS Credentials
-#[cfg(feature = "azure-kms")]
-#[tokio::test]
-async fn azure_imds() -> Result<()> {
-    let mut azure_exec = crate::client::csfle::state_machine::azure::ExecutorState::new()?;
-    azure_exec.test_host = Some(("localhost", *AZURE_IMDS_MOCK_PORT));
-
-    // Case 1: Success
-    {
-        let now = std::time::Instant::now();
-        let token = azure_exec.get_token().await?;
-        assert_eq!(token, rawdoc! { "accessToken": "magic-cookie" });
-        let cached = azure_exec.take_cached().await.expect("cached token");
-        assert_eq!(cached.server_response.expires_in, "70");
-        assert_eq!(cached.server_response.resource, "https://vault.azure.net");
-        assert!((65..75).contains(&cached.expire_time.duration_since(now).as_secs()));
-    }
-
-    // Case 2: Empty JSON
-    {
-        azure_exec.test_param = Some("case=empty-json");
-        let result = azure_exec.get_token().await;
-        assert!(result.is_err(), "expected err got {:?}", result);
-        assert!(result.unwrap_err().is_auth_error());
-    }
-
-    // Case 3: Bad JSON
-    {
-        azure_exec.test_param = Some("case=bad-json");
-        let result = azure_exec.get_token().await;
-        assert!(result.is_err(), "expected err got {:?}", result);
-        assert!(result.unwrap_err().is_auth_error());
-    }
-
-    // Case 4: HTTP 404
-    {
-        azure_exec.test_param = Some("case=404");
-        let result = azure_exec.get_token().await;
-        assert!(result.is_err(), "expected err got {:?}", result);
-        assert!(result.unwrap_err().is_auth_error());
-    }
-
-    // Case 5: HTTP 500
-    {
-        azure_exec.test_param = Some("case=500");
-        let result = azure_exec.get_token().await;
-        assert!(result.is_err(), "expected err got {:?}", result);
-        assert!(result.unwrap_err().is_auth_error());
-    }
-
-    // Case 6: Slow Response
-    {
-        azure_exec.test_param = Some("case=slow");
-        let result = azure_exec.get_token().await;
-        assert!(result.is_err(), "expected err got {:?}", result);
-        assert!(result.unwrap_err().is_auth_error());
-    }
-
-    Ok(())
-}
 
 // Prose test 19. Azure IMDS Credentials Integration Test (case 1: failure)
 #[cfg(feature = "azure-kms")]
 #[tokio::test]
 async fn azure_imds_integration_failure() -> Result<()> {
+    use mongocrypt::ctx::KmsProvider;
+
     let c = ClientEncryption::new(
         Client::for_test().await.into_client(),
         KV_NAMESPACE.clone(),
@@ -2205,176 +2047,6 @@ async fn range_explicit_encryption_defaults() -> Result<()> {
     assert!(payload.bytes.len() > payload_defaults.bytes.len());
 
     Ok(())
-}
-
-// Prose Test 24. KMS Retry Tests
-// using openssl causes errors after configuring a network failpoint
-// requires failpoint server to be running
-#[cfg(not(feature = "openssl-tls"))]
-#[tokio::test]
-async fn kms_retry_skip_local() {
-    use super::{AZURE_KMS, CSFLE_TLS_CERT_DIR, GCP_KMS};
-    use crate::{test::SERVERLESS, Namespace};
-    use reqwest::{Certificate, Client as HttpClient};
-    use std::path::PathBuf;
-
-    if *SERVERLESS {
-        log_uncaptured("skipping kms_retry on serverless");
-        return;
-    }
-
-    let endpoint = "127.0.0.1:9003";
-
-    let mut certificate_file_path = PathBuf::from(&*CSFLE_TLS_CERT_DIR);
-    certificate_file_path.push("ca.pem");
-    let certificate_file = std::fs::read(&certificate_file_path).unwrap();
-
-    let set_failpoint = |kind: &str, count: u8| {
-        // create a fresh client for each request to avoid hangs
-        let http_client = HttpClient::builder()
-            .add_root_certificate(Certificate::from_pem(&certificate_file).unwrap())
-            .build()
-            .unwrap();
-        let url = format!("https://localhost:9003/set_failpoint/{}", kind);
-        let body = format!("{{\"count\":{}}}", count);
-        http_client.post(url).body(body).send()
-    };
-
-    let aws_kms = AWS_KMS.clone();
-    let mut azure_kms = AZURE_KMS.clone();
-    azure_kms.1.insert("identityPlatformEndpoint", endpoint);
-    let mut gcp_kms = GCP_KMS.clone();
-    gcp_kms.1.insert("endpoint", endpoint);
-    let mut kms_providers = vec![aws_kms, azure_kms, gcp_kms];
-
-    let tls_options = get_client_options().await.tls_options();
-    for kms_provider in kms_providers.iter_mut() {
-        kms_provider.2 = tls_options.clone();
-    }
-
-    let key_vault_client = Client::for_test().await.into_client();
-    let client_encryption = ClientEncryption::new(
-        key_vault_client,
-        Namespace::new("keyvault", "datakeys"),
-        kms_providers,
-    )
-    .unwrap();
-
-    let aws_master_key = AwsMasterKey::builder()
-        .region("foo")
-        .key("bar")
-        .endpoint(endpoint.to_string())
-        .build();
-    let azure_master_key = AzureMasterKey::builder()
-        .key_vault_endpoint(endpoint)
-        .key_name("foo")
-        .build();
-    let gcp_master_key = GcpMasterKey::builder()
-        .project_id("foo")
-        .location("bar")
-        .key_ring("baz")
-        .key_name("qux")
-        .endpoint(endpoint.to_string())
-        .build();
-
-    // Case 1: createDataKey and encrypt with TCP retry
-
-    // AWS
-    set_failpoint("network", 1).await.unwrap();
-    let key_id = client_encryption
-        .create_data_key(aws_master_key.clone())
-        .await
-        .unwrap();
-    set_failpoint("network", 1).await.unwrap();
-    client_encryption
-        .encrypt(123, key_id, Algorithm::Deterministic)
-        .await
-        .unwrap();
-
-    // Azure
-    set_failpoint("network", 1).await.unwrap();
-    let key_id = client_encryption
-        .create_data_key(azure_master_key.clone())
-        .await
-        .unwrap();
-    set_failpoint("network", 1).await.unwrap();
-    client_encryption
-        .encrypt(123, key_id, Algorithm::Deterministic)
-        .await
-        .unwrap();
-
-    // GCP
-    set_failpoint("network", 1).await.unwrap();
-    let key_id = client_encryption
-        .create_data_key(gcp_master_key.clone())
-        .await
-        .unwrap();
-    set_failpoint("network", 1).await.unwrap();
-    client_encryption
-        .encrypt(123, key_id, Algorithm::Deterministic)
-        .await
-        .unwrap();
-
-    // Case 2: createDataKey and encrypt with HTTP retry
-
-    // AWS
-    set_failpoint("http", 1).await.unwrap();
-    let key_id = client_encryption
-        .create_data_key(aws_master_key.clone())
-        .await
-        .unwrap();
-    set_failpoint("http", 1).await.unwrap();
-    client_encryption
-        .encrypt(123, key_id, Algorithm::Deterministic)
-        .await
-        .unwrap();
-
-    // Azure
-    set_failpoint("http", 1).await.unwrap();
-    let key_id = client_encryption
-        .create_data_key(azure_master_key.clone())
-        .await
-        .unwrap();
-    set_failpoint("http", 1).await.unwrap();
-    client_encryption
-        .encrypt(123, key_id, Algorithm::Deterministic)
-        .await
-        .unwrap();
-
-    // GCP
-    set_failpoint("http", 1).await.unwrap();
-    let key_id = client_encryption
-        .create_data_key(gcp_master_key.clone())
-        .await
-        .unwrap();
-    set_failpoint("http", 1).await.unwrap();
-    client_encryption
-        .encrypt(123, key_id, Algorithm::Deterministic)
-        .await
-        .unwrap();
-
-    // Case 3: createDataKey fails after too many retries
-
-    // AWS
-    set_failpoint("network", 4).await.unwrap();
-    client_encryption
-        .create_data_key(aws_master_key)
-        .await
-        .unwrap_err();
-
-    // Azure
-    set_failpoint("network", 4).await.unwrap();
-    client_encryption
-        .create_data_key(azure_master_key)
-        .await
-        .unwrap_err();
-
-    // GCP
-    set_failpoint("network", 4).await.unwrap();
-    client_encryption
-        .create_data_key(gcp_master_key)
-        .await
-        .unwrap_err();
 }
 
 // FLE 2.0 Documentation Example
